Remove and ignore generated files, add bootstrap script
This removes the files generated by autotools from the repository and
ignores changes to them.
To generate them, the 'bootstrap' script is added. For recent versions
of autotools also 'autoreconf -i' can be used, it is only needed for
compatibility with old versions.
To include the autotools files in release tarballs, 'make dist' can be
used, so it is only needed to run the 'bootstrap' script for source
checkouts.